+++ src/connectors/impls/http/server.rs
@@ -48,8 +48,7 @@ pub(crate) struct Config {
     tls: Option<TLSServerConfig>,
     /// custom codecs mapping from mime_type to custom codec name
     /// e.g. for handling `application/json` with the `binary` codec, if desired
-    #[serde(default)]
-    custom_codecs: HashMap<String, String>,
+    custom_codecs: Option<HashMap<String, String>>,
 }
 
 impl ConfigImpl for Config {}
@@ -93,7 +92,11 @@ impl ConnectorBuilder for Builder {
             .as_ref()
             .map_or_else(|| HttpServer::DEFAULT_CODEC.to_string(), |c| c.name.clone());
         let inflight = Arc::default();
-        let codec_map = MimeCodecMap::with_overwrites(&config.custom_codecs);
+        let codec_map = if let Some(custom_codecs) = config.custom_codecs.clone() {
+            MimeCodecMap::from_custom(custom_codecs)
+        } else {
+            MimeCodecMap::new()
+        };
 
         Ok(Box::new(HttpServer {
             config,