+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Breaking Changes
 
 * Move `no_delay` config of `tcp_server` and `ws_server` to `socket_options.TCP_NODELAY`.
+* The HTTP connectors now work with the `mime_mapping` instead of `codec`
 
 ## [0.13.0-rc.9]